add compile method
